Make app settings available to text parsers
